Use Pickle as binary sources cache.
The cPickle serialization module is almost eight times
as fast as the generic YAML loader, and still almost
twice as fast as the optimized CLoader.

You have to run 'rosdep update' to rebuild the cache
in the binary format. If no binary cache files are
found, the code falls back to the old YAML
serialization.