allow non-existing system wide sources.list.d
Since #342 we have the option to override the source lists via ROSDEP_SOURCE_PATH.
This fix make it possible to use rosdep without any system wide sources.list.d (e.g. if you don't have sudo).

If ROSDEP_SOURCE_PATH is not set, nothing changes.
If it is set, the first directory in that list that exists is used (with fallback to system wide one in /etc).