Added a version checking class
Until now version were not used, but as we want to use them we must ensure of
their correctness.

Now if a dependency is provided with two version tags an error message will
be printed to the user and the version tag ignored.

If we have a same dependency used for mutiple purpose (for example build and
exec) if the version tag of both of them is irrealistic an error message will
be printed.

By irrealistic we mean this kind of situation:

<build_depend version_eq ="1">rospy</build_depend>
<exec_depend version_eq ="2">rospy</exec_depend>

or this:

<build_depend version_lt ="1">rospy</build_depend>
<exec_depend version_eq ="2">rospy</exec_depend>

On the other and here is an example of realistic situation that can be
statisfied:

<build_depend version_lt ="2">rospy</build_depend>
<exec_depend version_eq ="1">rospy</exec_depend>