Update the FreeBSD Installer
The FreeBSD system is now using pkg, which is modern, binary-based,
and comparable to apt or rpm. This starts using the system, with some
very rudimentary tests.

From discussion, the REP112 is going the way of the
dinosaur. So, the FreeBSD installer no longer uses this source installer.

If anyone wants to run the whole testsuite on FreeBSD, you'll also
need to make some stub programs for apt-get to work (a simple script
that just return zero will suffice), install rpm, and link bash into
/bin (sacrilige on a FreeBSD system). Note that that's just to get rid
of false positive for tests that run on other platforms.