Fix detection of installed rpms and warn if slow method is being used
The fallback method for detecting installed rpms (rpm_detect_cmd) was returning a list of all packages passed to it instead of only those installed.
On line 92, the .split() was breaking up an error, "no package provides <package>", that rpm outputs.
The list 'out_lines' then had an element for every package passed into the function, so the check on line 94 always returns true.

The fix is to split on newlines so that the error is a single element in the 'out_lines' list.
The string check on line 94 then fails on packages that aren't installed, as intended.

An error message was also added to warn the user that the fallback method is being used if debugging is enabled.