Fix error in rpm_expand_cmd
The check to see if there was a macro to be expanded was in rpm_expand, but rpm_detect_cmd calls rpm_expand_cmd directly. This fix moves the check from the rpm_expand function to both rpm_expand_cmd and rpm_expand_py. The check was moved to both functions rather than calling rpm_expand to avoid attempting to import rpm for every package to be checked.

It is currently unclear why calling rpm_expand_cmd on a non-macro string returns an empty string in python but not on the command line.