Fix detection of installed rpms and warn if slow method is being used (#568)
* Fix detection of installed rpms and warn if slow method is being used

The fallback method for detecting installed rpms (rpm_detect_cmd) was returning a list of all packages passed to it instead of only those installed.
On line 92, the .split() was breaking up an error, "no package provides <package>", that rpm outputs.
The list 'out_lines' then had an element for every package passed into the function, so the check on line 94 always returns true.

The fix is to split on newlines so that the error is a single element in the 'out_lines' list.
The string check on line 94 then fails on packages that aren't installed, as intended.

An error message was also added to warn the user that the fallback method is being used if debugging is enabled.

* Fix error in rpm_expand_cmd

The check to see if there was a macro to be expanded was in rpm_expand, but rpm_detect_cmd calls rpm_expand_cmd directly. This fix moves the check from the rpm_expand function to both rpm_expand_cmd and rpm_expand_py. The check was moved to both functions rather than calling rpm_expand to avoid attempting to import rpm for every package to be checked.

It is currently unclear why calling rpm_expand_cmd on a non-macro string returns an empty string in python but not on the command line.

* Expanded rpm tests and added real test data for fedora