Direct test for redhat.rpm_expand_cmd at the function itself
This test only applies to rpm_expand_cmd - if they RPM python module
is installed, that will be used instead, which makes this test break
on those systems (when %fedora doesn't expand to 27).

If we want to cover redhat.rpm_expand or redhat.rpm_expand_py, we
should devise a different approach.