Fix the unit test test_get_gbprepo_as_rosdep_data
The unit test checks for the existance of thre keys ['ros', 'catkin', 'genmsg'] in the rosdep data gather from a distro. As coded, it gets the first distribution found in alphabetical order. As of today "ardent". However this is a ros2 distribution and, for example, catkin is not released for it  (https://github.com/ros/rosdistro/blob/master/ardent/distribution.yaml)