Show rosdep-fix-permissions advice when cache read fails.
By re-raising this IOError as a CachePermissionsError the advice to run
rosdep fix-permissions will be displayed instead of the general
stacktrace.

If rosdep update is run as root and then a command like rosdep resolve
is used to read from the cache rather than write to it this stacktrace
occurs as reported in #782.