Refactor all_dep_types() function into constant.
Constants are used in rosdep and catkin_pkg for these kinds of data even
though there is a risk that they can be modified. In addition to
converting from a function to a constant I also relocated the constant
definition to the catkin_packages module in rosdep in order to keep it
closer to catkin_pkg related content.