Fetch upstream rosdistro data if present.
Recursively fetch rosdep data from upstream distributions.
Information from the rosdistro index file is required to check for the
presence of an upstream distribution. If present, rosdep data for that
upstream distribution will be fetched before processing the current
distribution.
Although it has not been tested, the recursive implementation should
allow for an arbitrary number of nested upstreams.