Pass with flake8_comprehensions
This flake8 plugin is installed as part of the ROS 2 development
packages, so it's pretty likely that anyone developing rosdep will
already have this plugin present. Since the plugin's mere presence will
cause it to load and run, it's probably easiest to just make this
package pass the checks.

They're valid issues, anyway.