Drop ROS Python package dependencies in debs
When our Python packages are split into a separate `-modules`
subpackage, it appears that the pattern is to retain the package
distribution information in the main package, which contains the
scripts.

When our packages take dependencies on only the `-modules` subpackage,
they need to omit the Python dependencies to prevent `pkg_resources`
from looking for those missing packages. Instead, we rely on apt and
dpkg to ensure that the dependencies are present.

This shouldn't have any affect on non-debian installation of rosdep.