Drop some over-agressive asserts from test_rosdep_main
These asserts check that nothing was written to stderr during some
top-level rosdep invocations. On derivative platforms like CentOS and
AlmaLinux, a message is written to stderr during normal operation
stating that the platform is an alias to the parent platform (RHEL, in
this example). This is normal and expected behavior on those platforms,
so the tests should still pass there.