Teach rosdep to use ROS_VERSION when resolving conditionals
The metadata from REP153 includes the ROS_PYTHON_VERSION, which is
currently cached during a call to `rosdep update` so that it is
available to package manifest conditionals during rosdep operations in
an environment where ROS_PYTHON_VERSION is not already set.

This change extends that metadata caching to include the ROS_VERSION
variable as well. These two values, combined with ROS_DISTRO, are the
three variables supplied by Bloom during manifest conditional
evaluation.