Check for pip configuration when an externally managed environment is detected. (#979)
* Raise an error when an externally managed environment is detected.

rosdep is designed to treat pip like an alternative system-level package
manager. Deviating from this approach is not easily achievable without a
significant rethinking of how pip packages are managed. In the meantime,
we can at least instruct users how to restore the prior functionality.

Rather than inject the environment variable / config on behalf of the
user, this change instructs them to make the necessary config changes
themself, keeping them informed of the change their making to the
system's new default.

* Fix typos and punctuation.

Co-authored-by: Christophe Bedard <bedard.christophe@gmail.com>

* Use tuple rather than list literal.

Co-authored-by: Christophe Bedard <bedard.christophe@gmail.com>

* Consolidate version check.

* Pass necessary environment variable via sudo.

* Update test to expect sudo --preserve-env for pip.

* flake8 cleanup

* Run pip tests with PIP_BREAK_SYSTEM_PACKAGES=1.

* Add documentation for pip configuration.

* Add doc link to error output.

* Use inline monospace font to refer to rosdep the cli tool.

* Briefly note that sudo configuration could prevent this from working.

* Recommend a specific config file to use and format user config as a warning.

* Fix errors in config checker.

The fallback configuration was over-indented and would never be
checked.

* Change formatting of rosdep.

Use monospace formatting when referring to the `rosdep` command /
executable name and simply 'rosdep' when referring to the project.

The preferred capitalization of rosdep is rosdep not Rosdep or ROSdep
(and certainly not ROSDep).

* Complete a sentence I stopped writing.

* Add period to end of sentence.

Co-authored-by: Christophe Bedard <bedard.christophe@gmail.com>

* Add period to end of sentence.

Co-authored-by: Christophe Bedard <bedard.christophe@gmail.com>

* Invert conditional for an earlier return.

Co-authored-by: Christophe Bedard <bedard.christophe@gmail.com>

* Edit text for clarity and typos.

Co-authored-by: Christophe Bedard <bedard.christophe@gmail.com>

* Reflow conditional for easier reading.

* Fix control flow after inverting the conditional.

This is a fixup after two earlier changes inverted the conditional and
reformatted an internal check.

* Add test to confirm that get_install_command handles externally managed environments.

* Use ConfigParser.getboolean to check config value.

Using the dict access method will raise a KeyError when the config file
is present but this section or value is missing.

---------

Co-authored-by: Christophe Bedard <bedard.christophe@gmail.com>