Add earthly build configuration for building docs.
I prefer the now acceptable build.earth filename over for the
Filefile-style Earthfile.

Along with the build.earth I've also included a .earthlyignore file
which uses the same format as a [.dockerignore] file.
Although we are not creating distributable images ignoring spurious
files in the docker context helps cut down on cache misses invalidating
dependency layers in the build.

[.dockerignore]: https://docs.docker.com/build/concepts/context/#dockerignore-files