Use rosdep's reported platform for testing OS override
This part of the test attempts to invoke a 'check' operation given a
specific OS override value. The test currently determines which OS:VER
pair to pass by directly calling into `rospkg.os_detect`. While this
works as expected for some platforms, it isn't appropriate for platforms
which either:
1. Use the OS_VERSION as the key version specifier as opposed to
   OS_CODENAME.
2. Implicitly override the OS_NAME to alias to another platform.

This change switches to using rosdep's own mechanism to determine which
test values to use. There should be no change in behavior to platforms
for which neither of the aforementioned cases apply to (e.x. Ubuntu,
Debian).