Switch to supported mechanism to skip an unwanted test (#995)
The 'git blame' for this return statement states that the test isn't
passing right now and references a GitHub ticket. While returning a
value from a test might be supported in other test frameworks, it is
deprecated in unittest. Rather than returning early from the test so
that it is reported as 'passed', we should use a supported mechanism to
skip the test so that it is reported as such.

Since the test is implemented using `unittest`, I used that mechanism to
skip the test, which also works as expected for pytest.