Remove unecessary global declarations.
Because these variables are in the local scope and are not re-assigned
in these functions, the global declaration can be omitted.

Recent versions of flake8 now raise F824[1] on these lines.
I am not sure if they were added defensively or if they were previously
required to access these variables as globals. Tests all pass as-is but
if we prefer to keep the defensive global declaration we can modify
these lines to `noqa: F824`.

[1]: https://flake8.pycqa.org/en/latest/user/error-codes.html